requirements.txt
attrs==25.3.0
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
looker-sdk==25.4.0
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
